Revert "Feature/lsh level3 4"
