+++ seokhwan/HarryPotterBooks/HarryPotterBooks/Presentation/MainViewController.swift
@@ -6,9 +6,32 @@ final class MainViewController: UIViewController {
     private let viewModel = MainViewModel()
     private var cancellables = Set<AnyCancellable>()
 
-    private lazy var headerView = HPBHeaderView()
-    private lazy var scrollView = UIScrollView()
-    private lazy var contentsView = HPBContainerView()
+    private lazy var bookTitleLabel: UILabel = {
+        let label = UILabel()
+
+        label.text = "HarryPotterBooks"
+        label.textAlignment = .center
+        label.font = .boldSystemFont(ofSize: 24)
+        label.numberOfLines = 0
+        label.lineBreakMode = .byWordWrapping
+
+        return label
+    }()
+    private lazy var seriesNumberButtonsStackView = UIStackView()
+    private lazy var seriesNumberButton: UIButton = {
+        var container = AttributeContainer()
+        container.font = UIFont.systemFont(ofSize: 16)
+
+        var configuration = UIButton.Configuration.filled()
+        configuration.attributedTitle = AttributedString("1", attributes: container)
+
+        let button = UIButton(configuration: configuration)
+        button.layer.cornerRadius = 22
+        button.layer.masksToBounds = true
+
+        return button
+    }()
+    private lazy var bookInformationView = BookInformationView()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -16,8 +39,17 @@ final class MainViewController: UIViewController {
     }
 
     private func updateContents(with book: Book) {
-        headerView.updateContents(with: book)
-        contentsView.updateContents(with: book)
+        bookTitleLabel.text = book.title
+        seriesNumberButton.configuration?.title = "\(book.seriesNumber)"
+        bookInformationView.updateContents(with: book)
+    }
+
+    private func presentErrorAlert(with message: String) {
+        let alert = UIAlertController(title: "Error", message: message, preferredStyle: .alert)
+        let okAction = UIAlertAction(title: "OK", style: .default)
+        alert.addAction(okAction)
+
+        present(alert, animated: true)
     }
 }
 
@@ -31,50 +63,48 @@ private extension MainViewController {
 
     func configureLayout() {
         view.backgroundColor = .systemBackground
-        scrollView.showsVerticalScrollIndicator = false
     }
 
     func configureSubviews() {
-        scrollView.addSubview(contentsView)
-        [headerView, scrollView].forEach {
+        [seriesNumberButton].forEach {
+            seriesNumberButtonsStackView.addArrangedSubview($0)
+        }
+        [bookTitleLabel, seriesNumberButtonsStackView, bookInformationView].forEach {
             view.addSubview($0)
         }
     }
 
     func configureConstraints() {
-        headerView.snp.makeConstraints { make in
+        bookTitleLabel.snp.makeConstraints { make in
             make.top.equalTo(view.safeAreaLayoutGuide).inset(10)
             make.directionalHorizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(20)
         }
-        scrollView.snp.makeConstraints { make in
-            make.top.equalTo(headerView.snp.bottom).offset(24)
-            make.directionalHorizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(20)
-            make.bottom.equalTo(view.safeAreaLayoutGuide)
+        seriesNumberButtonsStackView.snp.makeConstraints { make in
+            make.top.equalTo(bookTitleLabel.snp.bottom).offset(16)
+            make.centerX.equalTo(view.safeAreaLayoutGuide)
         }
-        contentsView.snp.makeConstraints { make in
-            make.directionalEdges.width.equalToSuperview()
+        seriesNumberButton.snp.makeConstraints { make in
+            make.size.equalTo(44)
+        }
+        bookInformationView.snp.makeConstraints { make in
+            make.top.equalTo(seriesNumberButtonsStackView.snp.bottom).offset(24)
+            make.directionalHorizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(5)
         }
     }
 
     func configureBind() {
-        viewModel.selectedBook
+        viewModel.$selectedBook
+            .receive(on: DispatchQueue.main)
             .compactMap { $0 }
             .sink { [weak self] book in
-                Task {
-                    await MainActor.run {
-                        self?.updateContents(with: book)
-                    }
-                }
+                self?.updateContents(with: book)
             }
             .store(in: &cancellables)
-        viewModel.errorMessage
+        viewModel.$errorMessage
+            .receive(on: DispatchQueue.main)
             .compactMap { $0 }
             .sink { [weak self] errorMessage in
-                Task {
-                    await MainActor.run {
-                        self?.presentErrorAlert(with: errorMessage)
-                    }
-                }
+                self?.presentErrorAlert(with: errorMessage)
             }
             .store(in: &cancellables)
     }
