+++ seokhwan/HarryPotterBooks/HarryPotterBooks/Presentation/MainViewModel.swift
@@ -1,39 +1,30 @@
 import Foundation
 import Combine
 
-protocol MainViewModelInput {
-    func selectBook(at index: Int)
-}
-
-protocol MainViewModelOutput {
-    var selectedBook: CurrentValueSubject<Book?, Never> { get }
-    var errorMessage: CurrentValueSubject<String?, Never> { get }
-}
-
-final class MainViewModel: MainViewModelInput, MainViewModelOutput {
+final class MainViewModel {
     private var books = Books()
+    @Published var selectedBook: Book?
+    @Published var errorMessage: String?
+
     private let fetchBooksUseCase: FetchableBooksUseCase
 
-    var selectedBook = CurrentValueSubject<Book?, Never>(nil)
-    var errorMessage = CurrentValueSubject<String?, Never>(nil)
-
-    init(fetchBooksUseCase: FetchableBooksUseCase = FetchBooksUseCase()) {
-        self.fetchBooksUseCase = fetchBooksUseCase
+    init() {
+        fetchBooksUseCase = FetchBooksUseCase()
         loadBooks()
     }
 
-    func selectBook(at index: Int) {
-        guard books.indices.contains(index) else { return }
-        selectedBook.send(books[index])
-    }
-
-    private func loadBooks() {
+    func loadBooks() {
         switch fetchBooksUseCase.execute() {
         case .failure(let error):
-            errorMessage.send(error.localizedDescription)
+            errorMessage = error.localizedDescription
         case .success(let books):
             self.books = books
             selectBook(at: books.startIndex)
         }
+    }
+
+    func selectBook(at index: Int) {
+        guard books.indices.contains(index) else { return }
+        selectedBook = books[index]
     }
 }
