+++ seokhwan/HarryPotterBooks/HarryPotterBooks/Storage/DataService.swift
@@ -3,15 +3,6 @@ import Foundation
 enum DataError: Error {
     case fileNotFound
     case parsingFailed
-
-    var localizedDescription: String {
-        switch self {
-        case .fileNotFound:
-            return "파일을 찾을 수 없습니다."
-        case .parsingFailed:
-            return "데이터 파싱에 실패했습니다."
-        }
-    }
 }
 
 protocol DataServiceProtocol {
