--- seokhwan/HarryPotterBooks/HarryPotterBooks/Utility/Array+Extensions.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension Array where Element == Chapter {
-    var joinedTitles: String {
-        return map { $0.title }.joined(separator: "\n")
-    }
-}
