--- seokhwan/HarryPotterBooks/HarryPotterBooks/Utility/UIViewController+Extensions.swift
@@ -1,11 +1,0 @@
-import UIKit
-
-extension UIViewController {
-    func presentErrorAlert(with message: String) {
-        let alert = UIAlertController(title: "오류", message: message, preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default)
-        alert.addAction(okAction)
-
-        present(alert, animated: true)
-    }
-}
