--- seokhwan/HarryPotterBooks/HarryPotterBooksTests/Mock/FetchBooksUseCaseMock.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-struct FetchBooksUseCaseMock: FetchableBooksUseCase {
-    var result: Result<Books, DataError>
-
-    func execute() -> Result<Books, DataError> {
-        return result
-    }
-}
