+++ suhyeon/HarryPotterBooks/HarryPotterBooks/Presentation/Home/Components/InfoStackView.swift
@@ -1,5 +1,5 @@
 //
-//  InfoHorizontalStackView.swift
+//  InfoStackView.swift
 //  HarryPotterBooks
 //
 //  Created by 이수현 on 3/26/25.
@@ -8,34 +8,34 @@
 import UIKit
 
 // 책 설명 수평 스택 뷰 
-class InfoHorizontalStackView: UIStackView {
+class InfoStackView: UIStackView {
     private let title: String
     private let titleFont: UIFont
     private let titleColor: UIColor
-    private let contentFont: UIFont
-    private let contentColor: UIColor
+    private let valueFont: UIFont
+    private let valueColor: UIColor
     
-    // 타이틀
+    // 저자 타이틀
     private lazy var titleLabel = UILabel().then {
         $0.text = title
         $0.font = titleFont
         $0.textColor = titleColor
     }
     
-    // 컨텐츠
-    private lazy var contentLabel = UILabel().then {
-        $0.font = contentFont
-        $0.textColor = contentColor
+    // 저자
+    private lazy var valueLabel = UILabel().then {
+        $0.font = valueFont
+        $0.textColor = valueColor
     }
     
     init(title: String, titleFont: UIFont, titleColor: UIColor,
-         contentFont: UIFont, contentColor: UIColor) {
+         valueFont: UIFont, valueColor: UIColor) {
         self.title = title
         self.titleFont = titleFont
         self.titleColor = titleColor
 
-        self.contentFont = contentFont
-        self.contentColor = contentColor
+        self.valueFont = valueFont
+        self.valueColor = valueColor
         super.init(frame: .zero)
     
         setSubview()
@@ -49,7 +49,7 @@ class InfoHorizontalStackView: UIStackView {
     private func setSubview(){
         [
             titleLabel,
-            contentLabel
+            valueLabel
         ].forEach { self.addArrangedSubview($0) }
     }
     
@@ -58,7 +58,7 @@ class InfoHorizontalStackView: UIStackView {
         self.spacing = 8
     }
     
-    func configure(content: String) {
-        contentLabel.text = content
+    func configure(value: String) {
+        valueLabel.text = value
     }
 }
