+++ ukhyun/HarryPotterNovelInfo/HarryPotterNovelInfo/ViewController.swift
@@ -52,11 +52,11 @@ class ViewController: UIViewController {
         let authorAttribute: [NSAttributedString.Key: Any] = [.font: UIFont.systemFont(ofSize: 18), .foregroundColor: UIColor.darkGray]
         let authorAttributeString = NSAttributedString(string: "J. K. Rowling", attributes: authorAttribute)
         
-        let combineAttribute = NSMutableAttributedString()
-        combineAttribute.append(titleAttributeString)
-        combineAttribute.append(authorAttributeString)
+        let combineAttrubute = NSMutableAttributedString()
+        combineAttrubute.append(titleAttributeString)
+        combineAttrubute.append(authorAttributeString)
         
-        author.attributedText = combineAttribute
+        author.attributedText = combineAttrubute
         
         return author
     }()
@@ -75,48 +75,6 @@ class ViewController: UIViewController {
         return pg
     }()
     
-    private let dedicationTitle = {
-        let dd = UILabel()
-        dd.text = "Dedication"
-        dd.font = .boldSystemFont(ofSize: 18)
-        return dd
-    }()
-    
-    private let dedicationDetail = {
-        let detail = UILabel()
-        detail.text = "Test Text"
-        detail.font = .systemFont(ofSize: 14)
-        detail.textColor = .darkGray
-        detail.numberOfLines = 0
-        return detail
-    }()
-    
-    private let summaryTitle = {
-        let title = UILabel()
-        title.text = "Summary"
-        title.font = .boldSystemFont(ofSize: 18)
-        return title
-    }()
-    
-    private let summaryDetail = {
-        let detail = UILabel()
-        detail.text = "Test Text"
-        detail.font = .systemFont(ofSize: 14)
-        detail.textColor = .darkGray
-        detail.numberOfLines = 0
-        return detail
-    }()
-    
-    private let scrollView = {
-        let scroll = UIScrollView()
-        scroll.showsVerticalScrollIndicator = true
-        scroll.showsHorizontalScrollIndicator = false
-        scroll.backgroundColor = .cyan
-        return scroll
-    }()
-    
-    private let containerView = UIView()
-    
     private let dataService = DataService()
     private var books = [BookAttributes]()
     
@@ -127,20 +85,14 @@ class ViewController: UIViewController {
         configureLayout()
         configureView()
     }
+    
+    
 }
 
 extension ViewController {
     func configureHierarchy() {
         view.backgroundColor = .white
-        
-        [titleLabel, buttonStackView, scrollView].forEach { view.addSubview($0)}
-        
-        scrollView.addSubview(containerView)
-        
-        [bookImageView, bookTitle, author, releaseDate, bookPage,
-         dedicationTitle, dedicationDetail, summaryTitle, summaryDetail]
-            .forEach { containerView.addSubview($0) }
-        
+        [titleLabel, buttonStackView, bookImageView, bookTitle, author, releaseDate, bookPage].forEach { view.addSubview($0) }
     }
     
     func configureLayout() {
@@ -153,60 +105,35 @@ extension ViewController {
             make.directionalHorizontalEdges.equalToSuperview().inset(20)
             make.height.equalTo(40)
         }
-        scrollView.snp.makeConstraints { make in
-            make.top.equalTo(buttonStackView.snp.bottom).offset(10)
-            make.directionalHorizontalEdges.equalToSuperview()
-            make.bottom.equalToSuperview()
-        }
-        containerView.snp.makeConstraints { make in
-            make.top.leading.trailing.bottom.equalTo(scrollView.contentLayoutGuide)
-            make.width.equalTo(scrollView.frameLayoutGuide)
-        }
         bookImageView.snp.makeConstraints { make in
-            make.top.equalToSuperview().offset(24)
-            make.leading.equalToSuperview().offset(20)
+            make.top.equalTo(buttonStackView.snp.bottom).offset(24)
+            make.leading.equalTo(view.safeAreaLayoutGuide).offset(5)
             make.width.equalTo(100)
-            make.height.equalTo(bookImageView.snp.width).multipliedBy(1.5)
+            make.height.equalTo(150)
         }
         bookTitle.snp.makeConstraints { make in
-            make.top.equalToSuperview().offset(28)
+            make.top.equalTo(buttonStackView.snp.bottom).offset(28)
             make.leading.equalTo(bookImageView.snp.trailing).offset(16)
-            make.trailing.equalToSuperview().inset(20)
+            make.trailing.equalToSuperview().inset(5)
         }
         author.snp.makeConstraints { make in
             make.top.equalTo(bookTitle.snp.bottom).offset(12)
             make.leading.equalTo(bookImageView.snp.trailing).offset(16)
-            make.trailing.equalToSuperview().inset(20)
+            make.trailing.equalToSuperview().inset(5)
         }
         releaseDate.snp.makeConstraints { make in
             make.top.equalTo(author.snp.bottom).offset(12)
             make.leading.equalTo(bookImageView.snp.trailing).offset(16)
-            make.trailing.equalToSuperview().inset(20)
+            make.trailing.equalToSuperview().inset(5)
         }
         bookPage.snp.makeConstraints { make in
             make.top.equalTo(releaseDate.snp.bottom).offset(12)
             make.leading.equalTo(bookImageView.snp.trailing).offset(16)
-            make.trailing.equalToSuperview().inset(20)
+            make.trailing.equalToSuperview().inset(5)
         }
-        dedicationTitle.snp.makeConstraints { make in
-            make.top.equalTo(bookImageView.snp.bottom).offset(24)
-            make.leading.equalToSuperview().offset(20)
-        }
-        dedicationDetail.snp.makeConstraints { make in
-            make.top.equalTo(dedicationTitle.snp.bottom).offset(8)
-            make.directionalHorizontalEdges.equalToSuperview().inset(20)
-        }
-        summaryTitle.snp.makeConstraints { make in
-            make.top.equalTo(dedicationDetail.snp.bottom).offset(24)
-            make.leading.equalToSuperview().offset(20)
-        }
-        summaryDetail.snp.makeConstraints { make in
-            make.top.equalTo(summaryTitle.snp.bottom).offset(8)
-            make.directionalHorizontalEdges.equalToSuperview().inset(20)
-            make.bottom.equalToSuperview().inset(20)
-        }
+        
     }
-
+    
     func configureView() {
         for i in 1...7 {
             let btn = UIButton()
@@ -215,11 +142,12 @@ extension ViewController {
             btn.backgroundColor = .systemBlue
             btn.tag = i
             btn.addTarget(self, action: #selector(seriesButtonClicked), for: .touchUpInside)
-            
-            btn.layer.cornerRadius = 20
-            btn.clipsToBounds = true
-            
             buttonStackView.addArrangedSubview(btn)
+            
+            DispatchQueue.main.async {
+                btn.layer.cornerRadius = btn.bounds.width / 2
+                btn.clipsToBounds = true
+            }
         }
     }
     
@@ -234,21 +162,17 @@ extension ViewController {
             let releaseDateFormatter = DateFormatter()
             releaseDateFormatter.dateFormat = "yyyy-MM-dd"
             guard let date = releaseDateFormatter.date(from: tag.releaseDate) else { return print("Release Date Error")}
-            releaseDate.text = date.dateFormatter()
-            dedicationDetail.text = tag.dedication
-            summaryDetail.text = tag.summary
+            releaseDate.text = date.DateToString()
             
         }
     }
-
+    
     func loadBooks() {
         dataService.loadBooks { [weak self] result in
             guard let self else { return }
             switch result {
-            case .success(let books):
-                self.books = books
-            case .failure(_):
-                showAlert(text: "오류", message: "확인")
+            case .success(let books): self.books = books
+            case .failure(let error): print("Error: \(error.localizedDescription)")
             }
         }
     }
